debug Model

+++ TrainModel.py
@@ -67,10 +67,10 @@ def EuclideanDist(a, b):
 
 def run_training():
     
-    sketchs_placeholder = tf.placeholder(tf.float32)
-    images_neg_placeholder = tf.placeholder(tf.float32)
-    images_pos_placeholder = tf.placeholder(tf.float32)
-    keep_prob = tf.placeholder(tf.float32)
+    sketchs_placeholder = tf.placeholder(tf.float64)
+    images_neg_placeholder = tf.placeholder(tf.float64)
+    images_pos_placeholder = tf.placeholder(tf.float64)
+    keep_prob = tf.placeholder(tf.float64)
 
     # Three Branch Net
     sketch_dense = SketchNet(sketchs_placeholder, _weights = image_weights, _biases = image_biases, dropout_prob = keep_prob)
